# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fikirden Ürüne Proje Vitrini API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /student/projects:
    get:
      operationId: listMyProjects
      tags: [student]
      summary: List current student's projects
      responses:
        "200":
          description: My projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Project"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createProject
      tags: [student]
      summary: Create a new project
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectInput"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /student/projects/{id}:
    patch:
      operationId: updateProject
      tags: [student]
      summary: Update a student's own project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProjectPatch"
      responses:
        "200":
          description: Updated project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Project"
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteProject
      tags: [student]
      summary: Delete a student's own project
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Not authorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    AiUsage:
      type: object
      required: [isUsed, tools, usageAreas, verificationMethod]
      properties:
        isUsed:
          type: boolean
        tools:
          type: array
          items:
            type: string
        usageAreas:
          type: array
          items:
            type: string
        verificationMethod:
          type: string
    SecurityAndEthics:
      type: object
      required: [noRealPersonalData, noSecretsInGithub, copyrightCompliant]
      properties:
        noRealPersonalData:
          type: boolean
        noSecretsInGithub:
          type: boolean
        copyrightCompliant:
          type: boolean
...
    Project:
      type: object
      required: [id, title, slug, category, city, stage, status, coverUrl, screenshotUrls, techStack, githubUrl, demoUrl, aiUsage, securityAndEthics, createdAt, updatedAt, studentId, studentName, problem, targetAudience, solution, futurePlans, adminFeedback]
      properties:
        id:
          type: integer
        studentId:
          type: integer
        studentName:
          type: string
        title:
          type: string
        slug:
          type: string
        category:
          type: string
          enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        stage:
          type: string
          enum: [MVP, Prototip]
        problem:
          type: string
        targetAudience:
          type: string
        solution:
          type: string
        futurePlans:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: ["string", "null"]
        demoUrl:
          type: ["string", "null"]
        coverUrl:
          type: ["string", "null"]
        screenshotUrls:
          type: array
          items:
            type: string
        aiUsage:
          $ref: "#/components/schemas/AiUsage"
        securityAndEthics:
          $ref: "#/components/schemas/SecurityAndEthics"
        status:
          type: string
          enum: [pending, approved, rejected, draft]
        adminFeedback:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
...
    ProjectInput:
      type: object
      required: [title, category, city, stage, problem, targetAudience, solution, futurePlans, techStack, aiUsage, securityAndEthics, status]
      properties:
        title:
          type: string
          minLength: 3
        category:
          type: string
          enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        stage:
          type: string
          enum: [MVP, Prototip]
        problem:
          type: string
        targetAudience:
          type: string
        solution:
          type: string
        futurePlans:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        demoUrl:
          type: string
        coverUrl:
          type: string
        screenshotUrls:
          type: array
          items:
            type: string
        aiUsage:
          $ref: "#/components/schemas/AiUsage"
        securityAndEthics:
          $ref: "#/components/schemas/SecurityAndEthics"
        status:
          type: string
          enum: [draft, pending]
    ProjectPatch:
      type: object
      properties:
        title:
          type: string
          minLength: 3
        category:
          type: string
          enum: [web-tasarim-ve-gelistirme, mobil-uygulama, oyun-gelistirme, yapay-zeka]
        city:
          type: string
          enum: [Konya, Afyonkarahisar]
        stage:
          type: string
          enum: [MVP, Prototip]
        problem:
          type: string
        targetAudience:
          type: string
        solution:
          type: string
        futurePlans:
          type: string
        techStack:
          type: array
          items:
            type: string
        githubUrl:
          type: string
        demoUrl:
          type: string
        coverUrl:
          type: string
        screenshotUrls:
          type: array
          items:
            type: string
        aiUsage:
          $ref: "#/components/schemas/AiUsage"
        securityAndEthics:
          $ref: "#/components/schemas/SecurityAndEthics"
        status:
          type: string
          enum: [draft, pending]
...