# lib/api-zod/src/generated/types/projectPatch.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Fikirden Ürüne Proje Vitrini API
 * OpenAPI spec version: 0.1.0
 */
import type { AiUsage } from './aiUsage';
import type { ProjectPatchCategory } from './projectPatchCategory';
import type { ProjectPatchCity } from './projectPatchCity';
import type { ProjectPatchStage } from './projectPatchStage';
import type { ProjectPatchStatus } from './projectPatchStatus';
import type { SecurityAndEthics } from './securityAndEthics';

export interface ProjectPatch {
  /** @minLength 3 */
  title?: string;
  category?: ProjectPatchCategory;
  city?: ProjectPatchCity;
  stage?: ProjectPatchStage;
  problem?: string;
  targetAudience?: string;
  solution?: string;
  futurePlans?: string;
  techStack?: string[];
  githubUrl?: string;
  demoUrl?: string;
  coverUrl?: string;
  screenshotUrls?: string[];
  aiUsage?: AiUsage;
  securityAndEthics?: SecurityAndEthics;
  status?: ProjectPatchStatus;
}